backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
pydantic==2.5.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# AI and ML capabilities
openai==1.3.8
anthropic==0.7.8
langchain==0.0.335
langchain-openai==0.0.2
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3

# Redis for caching and session management (Enterprise features)
redis==5.0.1
redis[hiredis]==5.0.1

# Background task processing
celery==5.3.4
flower==2.0.1

# WebSocket support for real-time features
websockets==12.0
fastapi-websocket-pubsub==0.3.0

# Enhanced security and auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Advanced monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Additional utilities
httpx==0.25.2
tenacity==8.2.3
email-validator==2.1.0
Pillow==10.1.0
python-slugify==8.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
